add static translation
It uses i18next and passes the t() function to the templates where
can be used as t('some string to translate'). If a string is not
found in the translation file, it is added to ns.missing.json,
where ns is the namespace.

We can use namespaces to keep the strings organized, for example,
the "navbar" namespace can contain all the strings related to the
navigation bar.

The current translation files are only for demonstration purposes.