Improve update of navigation menu
The method used in the last commit was being called in every page request,
which may be a performance issue because it fetch data from the Pages
collection every time. In this commit, the `updateNavigation` function and
the array of links were moved to the keystone.js file and are used only on
initialization and when the Pages collection is updated.